+++ src/test/scala/millfork/test/ConstantSuite.scala
@@ -13,7 +13,8 @@ class ConstantSuite extends FunSuite with Matchers {
       EmuUnoptimizedCrossPlatformRun(Cpu.Mos, Cpu.Z80, Cpu.Intel8086)(
         """
           | array Sieve[4]
-          | array __screen[4]
+          | const byte two = 2
+          | array __screen[4] = [4 / two, 4 %% two, 0, 0]
           | byte vic_mem
           | void main() {
           |   vic_mem = lo( ((Sieve.addr >> 10) & 8) | ((__screen.addr >> 6) & $f0) )